refactor: add unbatch in trainer
