+++ pyproject.toml
@@ -26,8 +26,8 @@ numpy = "^1.23.1"
 pandas = "^1.4.3"
 torch = "~1.12.0"
 scikit-learn = "^1.0"
-pytorch-lightning = "^1.8.3.post1"
 mlflow-skinny = { version = "~2.0.1", optional = true }
+pytorch-lightning = "^1.8.4.post0"
 
 [tool.poetry.extras]
 mlflow-skinny = ["mlflow-skinny"]
